Added stealth + observers
Added in my stealth system, on "E" press the player goes stealth for a certain amount of time, it plays a VFX and their materials alpha gets scaled down to make them transparent.

Also added in an observer system which is little floating robots that have a non stealth and stealth detection field that alarms if the player is inside them, i have a base observer class, then two children a stationary observer and a movingobserver both with different rules.